For #7: catch all exceptions from send_triage_reminder.
The root cause of the failure was that our credentials were expired and
unable to be refreshed automatically: I updated the credentials in the
Secrets to address that.

I also now catch all exceptions when sending the reminder to prevent the task
from failing to generate a rotation even if there is an error when
sending the reminder. It would be safer to separate the build tasks but
I didn't want to commit the time.

This push should also generate the rotation that failed and send the
appropriate calendar update.